% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_hg19-data.R, R/refseq_hg19-data.R
\docType{data}
\name{biomart_hg19}
\alias{biomart_hg19}
\title{Biomart Gene Data}
\format{A flat file}
\source{
\href{http://grch37.ensembl.org/index.html}

\href{http://grch37.ensembl.org/index.html}
}
\usage{
data(biomart_hg19)

data(biomart_hg19)
}
\description{
Data from the Biomart database for all known genes

Data from the Biomart database for all known genes
}
\examples{
data(biomart_hg19)
data(biomart_hg19)
}
\keyword{datasets}
