% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_hg38-data.R
\docType{data}
\name{biomart_hg38}
\alias{biomart_hg38}
\title{Biomart Gene Data}
\format{A flat file}
\source{
\href{http://useast.ensembl.org/biomart/martview}
}
\usage{
data(biomart_hg38)
}
\description{
Data from the Biomart database for all known genes
}
\examples{
data(biomart_hg38)
}
\keyword{datasets}
