% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19-data.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{Biomart Gene Data}
\format{A flat file}
\source{
\url{http://grch37.ensembl.org/index.html}
}
\usage{
data(hg19)
}
\description{
Data from the Biomart database for all known genes
}
\examples{
data(hg19)
}
\keyword{datasets}
