% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38-data.R
\docType{data}
\name{hg38}
\alias{hg38}
\title{Biomart Gene Data}
\format{A flat file}
\source{
\url{http://useast.ensembl.org/biomart/martview}
}
\usage{
data(hg38)
}
\description{
Data from the Biomart database for all known genes
}
\examples{
data(hg38)
}
\keyword{datasets}
