% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldRACER.R
\name{ldRACER}
\alias{ldRACER}
\title{Calculating Linkage Disequilibrium Information for Regional Association ComparER}
\usage{
ldRACER(assoc_data, rs_col, pops, lead_snp)
}
\arguments{
\item{assoc_data}{required. A dataframe produced by by formatRACER()}

\item{rs_col}{required. numeric. index of column containing rsID numbers for SNPs}

\item{pops}{required. Populations used to calculate LD.}

\item{lead_snp}{required. Required if ldby = "1000genomes". snp used to calculate LD}
}
\description{
This group of functions allows you to creat a plot of -log10(P-values) of an association study by their genomic position, for example, the results of a GWAS or eQTL study. This function takes the rsID of a reference SNP and calculates LD for all other SNPs in the dataset using the 1000 Genomes Phase III Data. The input of the function should already have been formatted using formatRACER().
}
\examples{
ldRACER(assoc_data = assoc_data, rs_col = 5, pops = c("POP1", "POP2", etc...), lead_snp = "RS_ID")
}
\keyword{association}
\keyword{disequilibrium.}
\keyword{gwas,}
\keyword{linkage}
\keyword{plot,}
