% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark3_bmd_gwas-data.R
\docType{data}
\name{mark3_bmd_gwas}
\alias{mark3_bmd_gwas}
\title{UKBB eBMD GWAS Data Release 2017}
\format{A flat file}
\source{
\url{http://www.gefos.org/?q=content/ukbb-ebmd-gwas-data-release-2017}
}
\usage{
data(mark3_bmd_gwas)
}
\description{
Summary statistics for the 2017 GWAS of eBMD from the UK BioBank in the region containing the gene Mark3
}
\examples{
data(mark3_bmd_gwas)
}
\keyword{datasets}
