% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark3_eqtl-data.R
\docType{data}
\name{mark3_eqtl}
\alias{mark3_eqtl}
\title{GTEx data for Mark3 from thyroid tissue}
\format{A flat file}
\source{
\url{https://gtexportal.org/home/}
}
\usage{
data(mark3_eqtl)
}
\description{
Data from the GTEx consortium, eQTL for Mark3 in thyroid
}
\examples{
data(mark3_eqtl)
}
\keyword{datasets}
