% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_plot.R
\name{mirror_plot_function}
\alias{mirror_plot_function}
\title{Mirror Plot Function}
\usage{
mirror_plot_function(assoc_data1, assoc_data2, name1, name2, x1, x2)
}
\arguments{
\item{assoc_data1}{object in your current R environment that has
columns named CHR_POS representing the position of the snp on
the chromosome, LOG10P which contains the -log10(P-values), and optional
columns CHR, which contains the chromosome number, RS_ID, which
contains the rsID numbers of the SNPs, and R2, which contains LD information.}

\item{assoc_data2}{identifcal in format to assoc_data1}

\item{name1}{name of association set 1}

\item{name2}{name of association set 2}

\item{x1}{lower bound on x axis}

\item{x2}{upper bound on x axis}
}
\description{
This function allows you to creat a plot of -log10(P-values).
}
\examples{
base_plot_function()
}
\keyword{association}
\keyword{plot}
