% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorPlotRACER.R
\name{mirrorPlotRACER}
\alias{mirrorPlotRACER}
\title{Mirror Plot -- Regional Association ComparER Plot}
\usage{
mirrorPlotRACER(assoc_data1, assoc_data2, chr, build = "hg19",
  name1 = "Association Dataset #1", name2 = "Association Dataset #2",
  plotby, gene_plot = NULL, snp_plot = NULL, start_plot = NULL,
  end_plot = NULL)
}
\arguments{
\item{assoc_data1}{required. A dataframe that has columns named
POS representing the position of the snp on
the chromosome, LOG10P which contains the -log10(P-values),
CHR, which contains the chromosome number, RS_ID, which
LD information. If no column named LD is in the input, LD will
be calculated from 1000 genomes phase III in relation to snp_ld_1}

\item{assoc_data2}{required. identifcal in format to assoc_data1}

\item{chr}{required. chromosome you wish to plot}

\item{build}{optional. indicates the genome build to use to plot the genes below the association plot. default = "hg19", but can be changed to "hg38"}

\item{name1}{optional. name of association set 1}

\item{name2}{optional. name of association set 2}

\item{plotby}{required. "coord", "gene", or "snp". Which parameter to use to
determine the reigon to be plotted.}

\item{gene_plot}{optional. If "gene" selected for plotby, then plot will be +/- 50kb of gene, should be a human gene symbol}

\item{snp_plot}{optional. If "snp" selected for plotby, then plot will be +/- 50kb of snp}

\item{start_plot}{optional. If "coord" selected for plotby, then this will be lower bound of x axis}

\item{end_plot}{optional. If "coord" selected for plotby, then this will be upper bound of x axis}
}
\description{
This function allows you to creat a plot of -log10(P-values).
}
\examples{
data(mark3_bmd_gwas)
data(mark3_eqtl)

mark3_bmd_gwas_f = RACER::formatRACER(assoc_data = mark3_bmd_gwas, chr_col = 3,
pos_col = 4, p_col = 11)
mark3_eqtl_f = RACER::formatRACER(assoc_data = mark3_eqtl, chr_col = 10,
pos_col = 11, p_col = 7)

mark3_bmd_gwas_f_ld = RACER::ldRACER(assoc_data = mark3_bmd_gwas_f,
rs_col = 2, pops ="EUR", lead_snp = "rs11623869")
mark3_eqtl_f_ld = RACER::ldRACER(assoc_data = mark3_eqtl_f,
rs_col = 15, pops = "EUR", lead_snp = "rs11623869")
mirrorPlotRACER(assoc_data1 = mark3_bmd_gwas_f_ld, assoc_data2 = mark3_eqtl_f_ld,
chr = 14, plotby = "gene", gene_plot = "MARK3")
}
\keyword{association}
\keyword{plot}
