% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_plot.R
\name{single_plot_function}
\alias{single_plot_function}
\title{Single Plot Function}
\usage{
single_plot_function(assoc_data)
}
\arguments{
\item{assoc_data}{a data frmae that has columns named CHR_POS representing
the position of the snp on the chromosome, LOG10P which contains the
-log10(P-values), and optional columns CHR, which contains the chromosome
number, RS_ID, which contains the rsID numbers of the SNPs, and R2, which
contains LD information.}
}
\description{
This function allows you to creat a plot of -log10(P-values).
}
\examples{
single_plot_function()
}
\keyword{association}
\keyword{plot}
