% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlePlotRACER.R
\name{singlePlotRACER}
\alias{singlePlotRACER}
\title{Single Regional Association Plot Function}
\usage{
singlePlotRACER(assoc_data, chr, build = "hg19", plotby,
  gene_plot = NULL, snp_plot = NULL, start_plot = NULL,
  end_plot = NULL)
}
\arguments{
\item{assoc_data}{required. A dataframe that has been produced by formatRACER and has columns named CHR, POS}

\item{chr}{required. numeric. chromosome to plot}

\item{build}{optional. default = "hg19", can also optionally be set to "hg38"}

\item{plotby}{required. "coord", "gene", or "snp". Which parameter to use to determine the reigon to be plotted.}

\item{gene_plot}{optional. Required if "gene" selected for plotby, then plot will be +/- 50kb of gene}

\item{snp_plot}{optional. Required if "snp" selected for plotby, then plot will be +/- 50kb of snp}

\item{start_plot}{optional. Required if "coord" selected for plotby, then this will be lower bound of x axis}

\item{end_plot}{optional. Required if "coord" selected for plotby, then this will be upper bound of x axis}
}
\description{
This function allows you to creat a plot of -log10(P-values) of an association study
by their genomic position, for example, the results of a GWAS or eQTL study. Sources
1000K genomes phase III data for linkage disequilibrium calculations.
}
\examples{
singlePlotRACER(assoc_data = assoc_data, chr = 1, plotby = "gene", gene_plot = "GENE_NAME")
singlePlotRACER(assoc_data = assoc_data, chr = 1, plotby = "snp", snp_plot = "RS_ID of lead SNP"
singlePlotRACER(assoc_data = assoc_data, chr = 1, plotby = "coord", start_plot = 100000, end_plot = 500000)
}
\keyword{association}
\keyword{disequilibrium.}
\keyword{gwas,}
\keyword{linkage}
\keyword{plot,}
