% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_plot.R
\name{mirror_plot_function}
\alias{mirror_plot_function}
\title{Mirror Plot Function}
\usage{
mirror_plot_function(assoc_data1, assoc_data2, chr, build = "hg19",
  name1 = "Association Dataset #1", name2 = "Association Dataset #2",
  plotby, gene_plot = NULL, snp_plot = NULL, start_plot = NULL,
  end_plot = NULL, ldby = "none", pops = NULL, snp_ld_1 = NULL,
  snp_ld_2 = NULL)
}
\arguments{
\item{assoc_data1}{required. A dataframe that has columns named
POS representing the position of the snp on
the chromosome, LOG10P which contains the -log10(P-values),
CHR, which contains the chromosome number, RS_ID, which
LD information. If no column named LD is in the input, LD will
be calculated from 1000 genomes phase III in relation to snp_ld_1}

\item{assoc_data2}{required. identifcal in format to assoc_data1}

\item{chr}{required. chromosome you wish to plot}

\item{name1}{optional. name of association set 1}

\item{name2}{optional. name of association set 2}

\item{plotby}{required. "coord", "gene", or "snp". Which parameter to use to
determine the reigon to be plotted.}

\item{gene_plot}{optional. If "gene" selected for plotby, then plot will be +/- 50kb of gene, should be a human gene symbol}

\item{snp_plot}{optional. If "snp" selected for plotby, then plot will be +/- 50kb of snp}

\item{start_plot}{optional. If "coord" selected for plotby, then this will be lower bound of x axis}

\item{end_plot}{optional. If "coord" selected for plotby, then this will be upper bound of x axis}
}
\description{
This function allows you to creat a plot of -log10(P-values).
}
\examples{
mirror_plot_function(assoc_data1 = df1, assoc_data2 = df2, chr = #,   name1 = "name1", name2 = "name2", plotby = "gene", gene_plot = "GENE_NAME")
mirror_plot_function(assoc_data1 = df1, assoc_data2 = df2, chr = #, name1 = "name1", name2 = "name2", plotby = "snp", snp_plot = "rs#######")
#' mirror_plot_function(assoc_data1 = df1, assoc_data2 = df2, chr = #, name1 = "name1", name2 = "name2", plotby = "coord", start_plot = #######, end_plot = #######)
}
\keyword{association}
\keyword{plot}
